chore(structures): cleanup
