+++ src/structures/Guild.ts
@@ -16,10 +16,10 @@ import { inverseActionType, parseAuditLog } from "./AuditLog.ts";
 import { Integration, parseIntegration } from "./Integration.ts";
 import { GuildEmoji, parseEmoji } from "./Emoji.ts";
 import type { StoreChannel } from "./StoreChannel.ts";
-import type { NewsChannel } from "./NewsChannel.ts";
+//import type { NewsChannel } from "./NewsChannel.ts";
 import type { CategoryChannel } from "./CategoryChannel.ts";
 import type { VoiceChannel } from "./VoiceChannel.ts";
-import type { TextChannel } from "./TextChannel.ts";
+import type { NewsChannel, TextChannel } from "./TextNewsChannel.ts";
 import { parseState, State } from "./VoiceState.ts";
 import { parsePresence, Presence } from "./Presence.ts";
 import { parseInvite, parseMetadataInvite } from "./Invite.ts";
