+++ src/structures/TextNewsChannel.ts
@@ -1,8 +1,6 @@
 import { EditOptions, TextBasedGuildChannel } from "./TextBasedGuildChannel.ts";
 import type { Client } from "../Client.ts";
-import type { channel } from "../discord.ts";
-import { Snowflake, webhook } from "../discord.ts";
-import type { NewsChannel } from "./NewsChannel.ts";
+import type { channel, Snowflake, webhook } from "../discord.ts";
 import { parseWebhook } from "./Webhook.ts";
 
 export class TextChannel<T extends channel.TextChannel = channel.TextChannel>
@@ -48,5 +46,37 @@ export class TextChannel<T extends channel.TextChannel = channel.TextChannel>
   async getWebhooks() {
     const webhooks = await this.client.rest.getChannelWebhooks(this.id);
     return webhooks.map((webhook) => parseWebhook(this.client, webhook));
+  }
+}
+
+export class NewsChannel<T extends channel.NewsChannel = channel.NewsChannel>
+  extends TextBasedGuildChannel<T> {
+  type = "news";
+
+  constructor(client: Client, data: T) {
+    super(client, data);
+  }
+
+  async edit(
+    options: EditOptions & { type?: "news" },
+    reason?: string,
+  ): Promise<NewsChannel>;
+  async edit(
+    options: EditOptions & { type: "text" },
+    reason?: string,
+  ): Promise<TextChannel>;
+  async edit(
+    options: EditOptions,
+    reason?: string,
+  ): Promise<TextChannel | NewsChannel> {
+    return super.edit(options, reason);
+  }
+
+  async delete(reason?: string) {
+    const channel = await this.client.rest.deleteChannel(
+      this.id,
+      reason,
+    ) as channel.NewsChannel;
+    return new NewsChannel(this.client, channel);
   }
 }
